Create separate class for LogIn action
